+++ plugins/aws/resources/provider.go
@@ -95,6 +95,7 @@ func Provider() *provider.Provider {
 			"rds.instances":                         RdsInstances(),
 			"redshift.clusters":                     RedshiftClusters(),
 			"redshift.subnet_groups":                RedshiftSubnetGroups(),
+			"route53.domains":                       Route53Domains(),
 			"route53.health_checks":                 Route53HealthChecks(),
 			"route53.hosted_zones":                  Route53HostedZones(),
 			"route53.reusable_delegation_sets":      Route53ReusableDelegationSets(),