+++ plugins/aws/CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed [#164](https://github.com/cloudquery/cq-provider-aws/issues/164) apigateway_api_keys does not return key value
 
 
+## [Unreleased]
+
+### :rocket: Added
+* Added Route53 Domains resource
+
 ## [v0.6.2] - 2021-11-03
 ###### SDK Version: 0.5.1